fixed viewer and extract
